requirements.txt
aniso8601==1.2.0
appdirs==1.4.3
asn1crypto==0.22.0
cffi==1.10.0
click==6.7
cryptography==1.8.1
enum34==1.1.6
Flask==0.12.1
Flask-Ask==0.8.9
Flask-RESTful==0.3.5
idna==2.5
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
packaging==16.8
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.2.0
pyserial==3.3
python-dateutil==2.6.0
pytz==2017.2
PyYAML==3.12
six==1.10.0
Werkzeug==0.12.1
